[REFACTOR]: 태그, 플레이 리스트 도메인으로 경로 수정
