+++ apps/api-user/src/main/java/com/ott/api_user/member/controller/MemberController.java
@@ -5,18 +5,13 @@ import com.ott.api_user.member.dto.request.UpdateMemberRequest;
 import com.ott.api_user.member.dto.response.*;
 import com.ott.api_user.member.service.MemberService;
 import com.ott.common.security.util.CookieUtil;
-import com.ott.common.web.response.PageResponse;
 import com.ott.common.web.response.SuccessResponse;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.Positive;
-import jakarta.validation.constraints.PositiveOrZero;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/member")
@@ -53,44 +48,6 @@ public class MemberController implements MemberApi {
         return ResponseEntity.noContent().build();
     }
 
-    // 유저 별 1달 간 상위 태그 조회
-    @Override
-    @GetMapping("/me/tag/ranking")
-    public ResponseEntity<SuccessResponse<TagRankingResponse>> getTagRanking(
-            @AuthenticationPrincipal Long memberId
-    ) {
-        return ResponseEntity.ok(SuccessResponse.of(memberService.getTagRanking(memberId)));
-    }
-
-    // 유저 별 2달 간 특정 태그 조회
-    @Override
-    @GetMapping("/me/tag/ranking/{tagId}")
-    public ResponseEntity<SuccessResponse<TagMonthlyCompareResponse>> getTagMonthlyCompare(
-            @AuthenticationPrincipal Long memberId,
-            @Positive @PathVariable Long tagId
-    ) {
-        return ResponseEntity.ok(SuccessResponse.of(memberService.getTagMonthlyCompare(memberId, tagId)));
-    }
-
-    // 태그 별 추천 리스트 조회
-    @Override
-    @GetMapping("/me/taglist/{tagId}")
-    public ResponseEntity<SuccessResponse<List<TagContentResponse>>> getRecommendContentsByTag(
-            @AuthenticationPrincipal Long memberId,
-            @Positive @PathVariable Long tagId
-    ) {
-        return ResponseEntity.ok(SuccessResponse.of(memberService.getRecommendContentsByTag(memberId, tagId)));
-    }
-
-    // 과거 시청 이력 조회, 10개씩 조회
-    @Override
-    @GetMapping("/me/history/playlist")
-    public ResponseEntity<SuccessResponse<PageResponse<RecentWatchResponse>>> getWatchHistoryPlaylist(
-            @AuthenticationPrincipal Long memberId,
-            @PositiveOrZero @RequestParam(defaultValue = "0") Integer page
-    ) {
-        return ResponseEntity.ok(SuccessResponse.of(memberService.getWatchHistoryPlaylist(memberId, page)));
-    }
 
     // 회원 탈퇴 - 현재 soft delete
     // 회원 탈퇴 시 DB + 브라우저 토큰 삭제
