+++ apps/api-user/src/main/java/com/ott/api_user/playlist/dto/response/RecentWatchResponse.java
@@ -1,4 +1,4 @@
-package com.ott.api_user.member.dto.response;
+package com.ott.api_user.playlist.dto.response;
 
 import com.ott.domain.common.MediaType;
 import com.ott.domain.watch_history.repository.RecentWatchProjection;
